+++ src/table.rs
@@ -13,7 +13,7 @@ impl<'a, T: Storage + 'a> Table<T> {
     /// where it should be inserted.
     pub fn find(&mut self, key: u32) -> Cursor<T> {
         let root_page_num = self.root_page_num;
-        let root_node = self.pager.page(root_page_num as usize);
+        let root_node = self.pager.page(root_page_num);
 
         match root_node {
             Node::Internal(internal) => internal.find(self, key),
@@ -58,22 +58,22 @@ impl<'a, T: Storage + 'a> Table<T> {
         let pager = &mut self.pager;
 
         // get old root page
-        let root = pager.page(self.root_page_num as usize);
+        let root = pager.page(self.root_page_num);
         let left_child_max_key = root.get_max_key();
 
         // get right child page
-        let mut right_child = pager.page(right_child_page_num as usize);
+        let mut right_child = pager.page(right_child_page_num);
 
         // get an unused page for the left child
         let left_child_page_num = pager.get_unused_page_num();
-        let mut left_child = pager.page(left_child_page_num as usize);
+        let mut left_child = pager.page(left_child_page_num);
 
         // Copy data from old root to left child
         left_child.buffer_mut().copy_from_slice(root.buffer());
         left_child.set_root(false);
 
         // Create a new root node as an internal node with one key and two children
-        let mut root = pager.new_internal_page(self.root_page_num as usize);
+        let mut root = pager.new_internal_page(self.root_page_num);
         root.node.set_root(true);
         root.set_num_keys(1);
         root.set_child(0, left_child_page_num);
@@ -109,7 +109,7 @@ impl<T: Storage> Cursor<T> {
             } else {
                 self.node = unsafe { &mut *self.table }
                     .pager
-                    .page(next_page_num as usize)
+                    .page(next_page_num)
                     .unwrap_leaf();
                 self.cell_num = 0;
             }
